Revert "rework VirshSSH.run method to better log/propagate errors"
This reverts commit f1b0801d5c0c4003fcc4025f3a700a692765cf72.

+++ src/provisioningserver/drivers/pod/virsh.py
@@ -316,13 +316,15 @@ class VirshSSH(pexpect.spawn):
 
     def _execute(self, poweraddr):
         """Spawns the pexpect command."""
-        self._spawn(f"virsh --connect {poweraddr}")
+        cmd = "virsh --connect %s" % poweraddr
+        self._spawn(cmd)
 
     def get_network_xml(self, network):
-        try:
-            return self.run(["net-dumpxml", network])
-        except VirshError:
+        output = self.run(["net-dumpxml", network]).strip()
+        if output.startswith("error:"):
+            maaslog.error("%s: Failed to get XML for network", network)
             return None
+        return output
 
     def get_machine_xml(self, machine):
         # Check if we have a cached version of the XML.
@@ -332,9 +334,9 @@ class VirshSSH(pexpect.spawn):
             return self.xml[machine]
 
         # Grab the XML from virsh if we don't have it already.
-        try:
-            output = self.run(["dumpxml", machine])
-        except VirshError:
+        output = self.run(["dumpxml", machine]).strip()
+        if output.startswith("error:"):
+            maaslog.error("%s: Failed to get XML for machine", machine)
             return None
 
         # Cache the XML, since we'll need it later to reconfigure the VM.
@@ -393,20 +395,12 @@ class VirshSSH(pexpect.spawn):
             return False
         return True
 
-    def run(self, args, raise_error=True):
+    def run(self, args):
         cmd = " ".join(args)
         self.sendline(cmd)
         self.prompt()
-        output = self.before.decode("utf-8").strip()
-        # remove the first line since it containes the issued command
-        output = "\n".join(output.splitlines()[1:])
-        if output.startswith("error:"):
-            message = f"Virsh command {args} failed: {output[7:]}"
-            maaslog.error(message)
-            if raise_error:
-                raise VirshError(message)
-            return ""  # return empty output if something failed
-        return output
+        result = self.before.decode("utf-8").splitlines()
+        return "\n".join(result[1:])
 
     def _get_column_values(self, data, keys):
         """Return tuple of column value tuples based off keys."""
@@ -448,14 +442,15 @@ class VirshSSH(pexpect.spawn):
             ["pool-autostart", "maas"],
         ]
         for command in commands:
-            try:
-                self.run(command)
-            except VirshError:
+            output = self.run(command)
+            if output.startswith("error:"):
+                maaslog.error("Failed to create Pod storage pool: %s", output)
                 return None
 
     def list_machines(self):
         """Lists all VMs by name."""
-        machines = self.run(["list", "--all", "--name"]).splitlines()
+        machines = self.run(["list", "--all", "--name"])
+        machines = machines.strip().splitlines()
         return [m for m in machines if m.startswith(self.dom_prefix)]
 
     def list_pools(self):
@@ -474,16 +469,16 @@ class VirshSSH(pexpect.spawn):
 
     def get_machine_state(self, machine):
         """Gets the VM state."""
-        try:
-            return self.run(["domstate", machine])
-        except VirshError:
+        state = self.run(["domstate", machine]).strip()
+        if state.startswith("error:"):
             return None
+        return state
 
     def get_machine_interface_info(self, machine):
         """Gets list of mac addressess assigned to the VM."""
-        try:
-            output = self.run(["domiflist", machine])
-        except VirshError:
+        output = self.run(["domiflist", machine]).strip()
+        if output.startswith("error:"):
+            maaslog.error("%s: Failed to get node MAC addresses", machine)
             return None
         return [
             InterfaceInfo(*entry)
@@ -502,7 +497,7 @@ class VirshSSH(pexpect.spawn):
 
     def get_machine_cpu_count(self, machine):
         """Gets the VM CPU count."""
-        output = self.run(["dominfo", machine])
+        output = self.run(["dominfo", machine]).strip()
         cpu_count = self.get_key_value(output, "CPU(s)")
         if cpu_count is None:
             maaslog.error("%s: Failed to get machine CPU count", machine)
@@ -528,7 +523,7 @@ class VirshSSH(pexpect.spawn):
 
     def get_machine_memory(self, machine):
         """Gets the VM memory."""
-        output = self.run(["dominfo", machine])
+        output = self.run(["dominfo", machine]).strip()
         KiB = self.get_key_value_unitless(output, "Max memory")
         if KiB is None:
             maaslog.error("%s: Failed to get machine memory", machine)
@@ -540,8 +535,8 @@ class VirshSSH(pexpect.spawn):
         """Get the storage pools information."""
         pools = []
         for pool in self.list_pools():
-            output = self.run(["pool-dumpxml", pool])
-            if not output:
+            output = self.run(["pool-dumpxml", pool]).strip()
+            if output is None:
                 # Skip if cannot get more information.
                 continue
 
@@ -589,8 +584,8 @@ class VirshSSH(pexpect.spawn):
 
     def get_machine_local_storage(self, machine, device):
         """Gets the VM local storage for device."""
-        output = self.run(["domblkinfo", machine, device])
-        if not output:
+        output = self.run(["domblkinfo", machine, device]).strip()
+        if output is None:
             maaslog.error("Failed to get available pod local storage")
             return None
         try:
@@ -600,7 +595,7 @@ class VirshSSH(pexpect.spawn):
 
     def get_pod_nodeinfo(self):
         """Gets the general information of the node via 'nodeinfo'"""
-        return self.run(["nodeinfo"])
+        return self.run(["nodeinfo"]).strip()
 
     def get_pod_arch(self, nodeinfo):
         """Gets architecture of the pod."""
@@ -769,22 +764,21 @@ class VirshSSH(pexpect.spawn):
         If no error is reported, destroy the domain to put it back into a
         'shut off' state.
         """
-        try:
-            self.run(["start", "--paused", machine])
-        except VirshError as e:
+        output = self.run(["start", "--paused", machine]).strip()
+        if output.startswith("error:"):
             # Delete the domain.
             self.delete_domain(machine)
             # Raise the error.
-            raise VirshError(f"Unable to compose {machine}: {e}")
+            raise VirshError("Unable to compose %s: %s" % (machine, output))
         else:
             # No errors, so set machine back to 'shut off' state.
             self.run(["destroy", machine])
 
     def set_machine_autostart(self, machine):
         """Set machine to autostart."""
-        try:
-            self.run(["autostart", machine])
-        except VirshError:
+        output = self.run(["autostart", machine]).strip()
+        if output.startswith("error:"):
+            maaslog.error("%s: Failed to set autostart", machine)
             return False
         return True
 
@@ -824,25 +818,24 @@ class VirshSSH(pexpect.spawn):
             f.write(etree.tostring(doc))
             f.write(b"\n")
             f.flush()
-            try:
-                self.run(["define", f.name])
-            except VirshError:
+            output = self.run(["define", f.name])
+            if output.startswith("error:"):
+                maaslog.error("%s: Failed to set network boot order", machine)
                 return False
+            maaslog.info("%s: Successfully set network boot order", machine)
             return True
 
     def poweron(self, machine):
         """Poweron a VM."""
-        try:
-            self.run(["start", machine])
-        except VirshError:
+        output = self.run(["start", machine]).strip()
+        if output.startswith("error:"):
             return False
         return True
 
     def poweroff(self, machine):
         """Poweroff a VM."""
-        try:
-            self.run(["destroy", machine])
-        except VirshError:
+        output = self.run(["destroy", machine]).strip()
+        if output.startswith("error:"):
             return False
         return True
 
@@ -942,7 +935,8 @@ class VirshSSH(pexpect.spawn):
 
     def get_volume_path(self, pool, volume):
         """Return the path to the file from `pool` and `volume`."""
-        return self.run(["vol-path", volume, "--pool", pool])
+        output = self.run(["vol-path", volume, "--pool", pool])
+        return output.strip()
 
     def attach_local_volume(self, domain, pool, volume, device):
         """Attach `volume` in `pool` to `domain` as `device`."""
@@ -966,7 +960,8 @@ class VirshSSH(pexpect.spawn):
 
     def get_network_list(self):
         """Return the list of available networks."""
-        return self.run(["net-list", "--name"]).splitlines()
+        output = self.run(["net-list", "--name"])
+        return output.strip().splitlines()
 
     def check_network_maas_dhcp_enabled(self, network, host_interfaces):
         xml = self.get_network_xml(network)
@@ -1104,10 +1099,17 @@ class VirshSSH(pexpect.spawn):
             f.write(device_xml.encode("utf-8"))
             f.write(b"\n")
             f.flush()
-            try:
-                self.run(["attach-device", domain, f.name, "--config"])
-            except VirshError:
+            output = self.run(["attach-device", domain, f.name, "--config"])
+            if output.startswith("error:"):
+                maaslog.error(
+                    "%s: Failed to attach network device %s"
+                    % (domain, interface.attach_name)
+                )
                 return False
+            maaslog.info(
+                "%s: Successfully attached network device %s"
+                % (domain, interface.attach_name)
+            )
             return True
 
     def get_domain_capabilities(self):
@@ -1116,13 +1118,23 @@ class VirshSSH(pexpect.spawn):
         Determines the type and emulator of the domain to use.
         """
         # Test for KVM support first.
-        emulator_type = "kvm"
-        try:
-            xml = self.run(["domcapabilities", "--virttype", emulator_type])
-        except VirshError:
+        xml = self.run(["domcapabilities", "--virttype", "kvm"])
+        if xml.startswith("error"):
             # Fallback to qemu support. Fail if qemu not supported.
+            xml = self.run(["domcapabilities", "--virttype", "qemu"])
             emulator_type = "qemu"
-            xml = self.run(["domcapabilities", "--virttype", emulator_type])
+        else:
+            emulator_type = "kvm"
+
+        # XXX newell 2017-05-18 bug=1690781
+        # Check to see if the XML output was an error.
+        # See bug for details about why and how this can occur.
+        if xml.startswith("error"):
+            raise VirshError(
+                "`virsh domcapabilities --virttype %s` errored.  Please "
+                "verify that package qemu-kvm is installed and restart "
+                "libvirt-bin service." % emulator_type
+            )
 
         doc = etree.XML(xml)
         evaluator = etree.XPathEvaluator(doc)
@@ -1245,9 +1257,8 @@ class VirshSSH(pexpect.spawn):
 
     def delete_domain(self, domain):
         """Delete `domain` and its volumes."""
-        # Ensure that its destroyed first. We ignore errors here not to fail
-        # the process, and since there isn't much we can do anyway.
-        self.run(["destroy", domain], raise_error=False)
+        # Ensure that its destroyed first.
+        self.run(["destroy", domain])
         # Undefine the domains and remove all storage and snapshots.
         # XXX newell 2018-02-25 bug=1741165
         # Removed the --delete-snapshots flag to workaround the volumes not
@@ -1259,8 +1270,7 @@ class VirshSSH(pexpect.spawn):
                 "--remove-all-storage",
                 "--managed-save",
                 "--nvram",
-            ],
-            raise_error=False,
+            ]
         )
